Build a Darwin Core Archive from a folder

A Darwin Core archive is a zip file containing a specified combination of data and metadata. build_archive() constructs this zip file. It assumes that all necessary files have been pre-constructed, and can be found inside a single folder with no additional or redundant information. Structurally, build_archive() is similar to devtools::build(), in the sense that it takes a repository and wraps it for publication.

Usage

build_archive(destination = "dwc-archive.zip")

Arguments

destination

(string) A file name to save the resulting zip file. Defaults to ./dwc-archive.zip.

Value

Invisibly returns the location of the built zip file; but typically called for the side-effect of building a 'Darwin Core Archive' (i.e. a zip file).

Details

This function looks for three types of objects in the specified directory:

Data

One or more csv files named occurrences.csv, events.csv and/or multimedia.csv. These csv files contain data standardised using Darwin Core Standard (see corella::corella-package() for details). A data.frame/tibble can be added to the correct folder using use_data().

Metadata

A metadata statement in EML format with the file name eml.xml. Completed metadata statements written markdown as .Rmd or qmd files can be converted and saved to the correct folder using use_metadata(). Create a new template with use_metadata_template().

Schema

A 'schema' document in xml format with the file name meta.xml. This file can be constructed using use_schema().

build_archive() will not build a Darwin Core Archive with these files present in the source directory. The resulting Archive is saved as a zip folder in the parent directory by default.

See also