Check whether a Darwin Core Archive meets Darwin Core Standard via API

Check whether a specified Darwin Core Archive is ready for sharing and publication, according to the Darwin Core Standard. check_archive() tests the specified archive using an online validation service by sending the archive via an API and returning the results. Currently only supports validation using GBIF.

Usage

check_archive(file = NULL, provider = "GBIF")

get_report(key)

Arguments

file

(string) A file name (ending in .zip) of the specified Darwin Core Archive.

provider

(string) The institution to be queried for validation services. Currently only "GBIF" is supported.

key

A unique identifier key from GBIF

Value

Invisibly returns a tibble to the workspace containing validation results; but primarily called for the side-effect of generating a report in the console.

Details

Results returned by check_archive() are directly from the institution API, not from galaxias.

See also

Examples

if (FALSE) { # \dontrun{
# add GBIF login details
galaxias_config(gbif = list(username = "your-gbif-username",
                            email = "your-gbif-email",
                            password = "your-gbif-password"))

# Check archive against Darwin Core Standard criteria
check_archive("dwc-archive.zip")
} # }