Create an example metadata statement file

This function creates a metadata template and exports it to the specified file. The template is taken from metadata_example (see metadata_example for documentation). The template .md file can act as a boilerplate metadata statement that you can edit yourself.

The final edited file can be converted for use in a Darwin Core Archive using build_metadata()

Usage

use_metadata(file, overwrite = FALSE)

Arguments

file

(string) A filename to save the statement to. Defaults to "metadata.md".

overwrite

(logical) Should any existing file with this name be overwritten? Defaults to FALSE.

Value

Does not return an object to the workspace; called for the side-effect of saving a markdown file to the specified location.

Examples

if (FALSE) { # \dontrun{
# Save template file metadata.md to local directory
use_metadata()
} # }