Create a schema for a Darwin Core Archive

A schema is an xml document that maps the files and field names in a DwCA. This map makes it easier to reconstruct one or more related datasets so that information is matched correctly. It works by detecting column names on csv files in a specified directory; these should all be Darwin Core terms for this function to produce reliable results.

Usage

use_schema(source = "data-publish", destination = "meta.xml")

Arguments

source

A folder containing one or more data csv files occurrences.csv, events.csv or multimedia.csv. Defaults to data-publish/.

destination

A file name for the resulting schema document. Defaults to meta.xml for consistency with the Darwin Core standard. Note this file is placed within the working directory specified by galaxias_config().

Value

Does not return an object to the workspace; called for the side effect of building a schema file in the specified directory.