Validate a Darwin Core Archive via API

Validation is a process of checking whether a specified archive is ready for sharing and publication, according to the Darwin Core standard.

Usage

validate_archive(x = "data", provider = "GBIF")

get_validator_report(key)

Arguments

x

(string) Either a directory containing all the files to be stored in the archive, or a filename (ending in .zip) of the specified archive. Defaults to the data folder within the current working directory.

provider

(string) The institution to be queried for validation services. Currently only "GBIF" is supported.

key

A unique identifier key from GBIF

Value

Invisibly returns a tibble to the workspace containing validation results; but primarily called for the side-effect of generating a report in the console.

See also