Create a schema for a Darwin Core Archive

A schema is an xml document that maps the files and field names in a DwCA. This map makes it easier to reconstruct one or more related datasets so that information is matched correctly. It works by detecting column names on csv files in a specified directory; these should all be Darwin Core terms for this function to produce reliable results. The publishing directory is set using galaxias_config() and defaults to "data-publish".

Usage

use_schema(overwrite = FALSE, quiet = FALSE)

Arguments

overwrite

By default, use_schema() will not overwrite existing files. If you really want to do so, set this to TRUE.

quiet

(logical) Should progress messages be suppressed? Default is set to FALSE; i.e. messages are shown.

Value

Does not return an object to the workspace; called for the side effect of building a schema file in the publication directory.

Details

To be compliant with the Darwin Core Standard, the schema file must be called meta.xml, and this function enforces that.

Examples

use_schema()
#> ℹ Building schema
#> Error in proj_set("."): ✖ Path /tmp/Rtmp9YEqun/ does not appear to be inside a project or
#>   package.
#> ℹ Read more in the help for `usethis::proj_get()`.